+++ apps/user-profile/src/components/modules/OnboardingLandingPage/index.js
@@ -1,0 +1,1 @@
+export { default } from "./OnboardingLandingPage";
