--- apps/user-profile/src/components/modules/UserOnboarding/index.js
@@ -1,1 +1,0 @@
-export { default } from "./UserOnboarding";
